requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
openai